# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Ledgerly accounting API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /vendors:
    get:
      operationId: listVendors
      tags: [vendors]
      summary: List vendors
      responses:
        "200":
          description: Vendors list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Vendor"
    post:
      operationId: createVendor
      tags: [vendors]
      summary: Create vendor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VendorInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
  /vendors/{id}:
    get:
      operationId: getVendor
      tags: [vendors]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Vendor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
    patch:
      operationId: updateVendor
      tags: [vendors]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VendorInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
    delete:
      operationId: deleteVendor
      tags: [vendors]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "204":
          description: Deleted
...
  /invoices:
    get:
      operationId: listInvoices
      tags: [invoices]
      summary: List invoices
      responses:
        "200":
          description: Invoices list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Invoice"
    post:
      operationId: createInvoice
      tags: [invoices]
      summary: Create invoice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
  /invoices/{id}:
    get:
      operationId: getInvoice
      tags: [invoices]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Invoice with line items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvoiceDetail"
    patch:
      operationId: updateInvoice
      tags: [invoices]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
    delete:
      operationId: deleteInvoice
      tags: [invoices]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "204":
          description: Deleted
  /invoices/{id}/void:
    post:
      operationId: voidInvoice
      tags: [invoices]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Voided
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
...
  /payments:
    get:
      operationId: listPayments
      tags: [payments]
      summary: List payments
      responses:
        "200":
          description: Payments list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Payment"
    post:
      operationId: createPayment
      tags: [payments]
      summary: Record payment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
  /payments/{id}:
    get:
      operationId: getPayment
      tags: [payments]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Payment detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
    delete:
      operationId: deletePayment
      tags: [payments]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Vendor:
      type: object
      required: [id, organizationId, name, createdAt]
      properties:
        id: { type: string }
        organizationId: { type: string }
        name: { type: string }
        email: { type: ["string", "null"] }
        phone: { type: ["string", "null"] }
        address: { type: ["string", "null"] }
        createdAt: { type: string }
    VendorInput:
      type: object
      required: [name]
      properties:
        name: { type: string }
        email: { type: ["string", "null"] }
        phone: { type: ["string", "null"] }
        address: { type: ["string", "null"] }
...
    LineItemInput:
      type: object
      required: [description, quantity, unitPriceCents, accountId]
      properties:
        description: { type: string }
        quantity: { type: integer }
        unitPriceCents: { type: integer }
        accountId: { type: string }
        taxRateId: { type: ["string", "null"] }
    DocumentInput:
      type: object
      required: [contactId, number, issueDate, dueDate, lines]
      properties:
        contactId: { type: string }
        number: { type: string }
        issueDate: { type: string }
        dueDate: { type: string }
        lines:
          type: array
          items:
            $ref: "#/components/schemas/LineItemInput"
    LineItem:
      type: object
      required: [id, description, quantity, unitPriceCents, accountId, amountCents, sortOrder]
      properties:
        id: { type: string }
        description: { type: string }
        quantity: { type: integer }
        unitPriceCents: { type: integer }
        accountId: { type: string }
        taxRateId: { type: ["string", "null"] }
        amountCents: { type: integer }
        sortOrder: { type: integer }
    Invoice:
      type: object
      required: [id, organizationId, customerId, number, status, issueDate, dueDate, subtotalCents, taxCents, totalCents, balanceCents, createdAt]
      properties:
        id: { type: string }
        organizationId: { type: string }
        customerId: { type: string }
        customerName: { type: ["string", "null"] }
        number: { type: string }
        status: { type: string }
        issueDate: { type: string }
        dueDate: { type: string }
        subtotalCents: { type: integer }
        taxCents: { type: integer }
        totalCents: { type: integer }
        balanceCents: { type: integer }
        createdAt: { type: string }
    InvoiceDetail:
      type: object
      required: [id, organizationId, customerId, number, status, issueDate, dueDate, subtotalCents, taxCents, totalCents, balanceCents, createdAt, lineItems]
      properties:
        id: { type: string }
        organizationId: { type: string }
        customerId: { type: string }
        customerName: { type: ["string", "null"] }
        number: { type: string }
        status: { type: string }
        issueDate: { type: string }
        dueDate: { type: string }
        subtotalCents: { type: integer }
        taxCents: { type: integer }
        totalCents: { type: integer }
        balanceCents: { type: integer }
        createdAt: { type: string }
        lineItems:
          type: array
          items:
            $ref: "#/components/schemas/LineItem"
...
    PaymentInput:
      type: object
      required: [direction, amountCents, date]
      properties:
        direction: { type: string }
        customerId: { type: ["string", "null"] }
        vendorId: { type: ["string", "null"] }
        amountCents: { type: integer }
        date: { type: string }
        method: { type: string }
        memo: { type: ["string", "null"] }
        allocations:
          type: array
          items:
            $ref: "#/components/schemas/AllocationInput"
    AllocationInput:
      type: object
      required: [amountCents]
      properties:
        invoiceId: { type: ["string", "null"] }
        billId: { type: ["string", "null"] }
        amountCents: { type: integer }
    Payment:
      type: object
      required: [id, organizationId, direction, amountCents, date, method, createdAt]
      properties:
        id: { type: string }
        organizationId: { type: string }
        direction: { type: string }
        customerId: { type: ["string", "null"] }
        customerName: { type: ["string", "null"] }
        vendorId: { type: ["string", "null"] }
        vendorName: { type: ["string", "null"] }
        amountCents: { type: integer }
        date: { type: string }
        method: { type: string }
        memo: { type: ["string", "null"] }
        createdAt: { type: string }
...