# lib/api-zod/src/generated/types/invoiceDetail.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Ledgerly accounting API
 * OpenAPI spec version: 0.1.0
 */
import type { LineItem } from './lineItem';

export interface InvoiceDetail {
  id: string;
  organizationId: string;
  customerId: string;
  /** @nullable */
  customerName?: string | null;
  number: string;
  status: string;
  issueDate: string;
  dueDate: string;
  subtotalCents: number;
  taxCents: number;
  totalCents: number;
  balanceCents: number;
  createdAt: string;
  lineItems: LineItem[];
}